Move theme-related assets to a theme folder
